test(native): cover resident cleanup ordering
